FIX: Visual Studio 2010 compilation issues
git-svn-id: svn://svn.code.sf.net/p/riccreator/code@163 dab27aee-19fe-a441-870f-52ce2490c35b

+++ src/GUI/nxtCanvasWidget.cpp
@@ -77,7 +77,7 @@ void nxtCanvasWidget::change_canvas( nxtCanvas* new_canvas, bool delete_old ){
 
 
 QPoint nxtCanvasWidget::point_to_pos( QPoint pos ) const{
-	QPoint delta = ( QPoint( pos_x, pos_y ) + pos ) * current_zoom;
+	QPoint delta = ( QPoint( pos_x, pos_y ) + pos ) * (int)current_zoom;
 	return QPoint( delta.x(), height()-delta.y() );
 }
